test(spec): shard out tests for rr (#1317)
